+++ twocatz/index.js
@@ -1,0 +1,17 @@
+const fetch = require('node-fetch')
+
+module.exports = async function (context, req) {
+    context.log('JavaScript HTTP trigger function processed a request.');
+
+    const resp = await fetch("https://bit-cat.azurewebsites.net/cat/says/serverless", {
+        method: 'GET'
+    });
+    
+    const data = await resp.arrayBuffer()
+
+    var base64data = Buffer.from(data).toString('base64')
+
+    context.res = {
+        body: { base64data }
+    }
+}
